mcp-toolbox/agent/BQdatalyzer/pyproject.toml
[project]
name = "bqdatalyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "aiosqlite==0.22.1",
    "alembic==1.18.4",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "attrs==25.4.0",
    "authlib==1.6.7",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "cloudpickle==3.1.2",
    "cryptography==46.0.5",
    "deprecated==1.3.1",
    "distro==1.9.0",
    "docstring-parser==0.17.0",
    "fastapi==0.129.0",
    "frozenlist==1.8.0",
    "google-adk==1.25.0",
    "google-api-core==2.29.0",
    "google-api-python-client==2.190.0",
    "google-auth==2.48.0",
    "google-auth-httplib2==0.3.0",
    "google-cloud-aiplatform==1.137.0",
    "google-cloud-appengine-logging==1.8.0",
    "google-cloud-audit-log==0.4.0",
    "google-cloud-bigquery==3.40.0",
    "google-cloud-bigquery-storage==2.36.1",
    "google-cloud-bigtable==2.35.0",
    "google-cloud-core==2.5.0",
    "google-cloud-discoveryengine==0.13.12",
    "google-cloud-iam==2.21.0",
    "google-cloud-logging==3.13.0",
    "google-cloud-monitoring==2.29.1",
    "google-cloud-pubsub==2.35.0",
    "google-cloud-resource-manager==1.16.0",
    "google-cloud-secret-manager==2.26.0",
    "google-cloud-spanner==3.62.0",
    "google-cloud-speech==2.36.1",
    "google-cloud-storage==3.9.0",
    "google-cloud-trace==1.18.0",
    "google-crc32c==1.8.0",
    "google-genai==1.63.0",
    "google-resumable-media==2.8.0",
    "googleapis-common-protos==1.72.0",
    "graphviz==0.21",
    "greenlet==3.3.1",
    "grpc-google-iam-v1==0.14.3",
    "grpc-interceptor==0.15.4",
    "grpcio==1.78.0",
    "grpcio-status==1.78.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.31.2",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "mako==1.3.10",
    "markupsafe==3.0.3",
    "mcp==1.26.0",
    "mmh3==5.2.0",
    "multidict==6.7.1",
    "opentelemetry-api==1.38.0",
    "opentelemetry-exporter-gcp-logging==1.11.0a0",
    "opentelemetry-exporter-gcp-monitoring==1.11.0a0",
    "opentelemetry-exporter-gcp-trace==1.11.0",
    "opentelemetry-exporter-otlp-proto-common==1.38.0",
    "opentelemetry-exporter-otlp-proto-http==1.38.0",
    "opentelemetry-proto==1.38.0",
    "opentelemetry-resourcedetector-gcp==1.11.0a0",
    "opentelemetry-sdk==1.38.0",
    "opentelemetry-semantic-conventions==0.59b0",
    "packaging==26.0",
    "propcache==0.4.1",
    "proto-plus==1.27.1",
    "protobuf==6.33.5",
    "pyarrow==23.0.0",
    "pyasn1==0.6.2",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pyjwt==2.11.0",
    "pyopenssl==25.3.0",
    "pyparsing==3.3.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.22",
    "pyyaml==6.0.3",
    "referencing==0.37.0",
    "requests==2.32.5",
    "rpds-py==0.30.0",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.46",
    "sqlalchemy-spanner==1.17.2",
    "sqlparse==0.5.5",
    "sse-starlette==3.2.0",
    "starlette==0.52.1",
    "tenacity==9.1.4",
    "toolbox-core==0.5.8",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzlocal==5.3.1",
    "uritemplate==4.2.0",
    "urllib3==2.6.3",
    "uvicorn==0.40.0",
    "vertexai==1.43.0",
    "watchdog==6.0.0",
    "websockets==15.0.1",
    "wrapt==2.1.1",
    "yarl==1.22.0",
    "zipp==3.23.0",
]
